--- be/issue/src/main/java/codesquad/issueTracker/global/StatusCode.java
@@ -1,10 +1,0 @@
-package codesquad.issueTracker.global;
-
-import org.springframework.http.HttpStatus;
-
-public interface StatusCode {
-
-    HttpStatus getStatus();
-
-    String getMessage();
-}
