+++ be/issue/src/main/java/codesquad/issueTracker/global/exception/StatusCode.java
@@ -1,0 +1,10 @@
+package codesquad.issueTracker.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface StatusCode {
+
+	HttpStatus getStatus();
+
+	String getMessage();
+}
