+++ be/issue/src/main/java/codesquad/issueTracker/global/exception/SuccessCode.java
@@ -1,0 +1,26 @@
+package codesquad.issueTracker.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum SuccessCode implements StatusCode {
+
+	SUCCESS(HttpStatus.OK, "요청이 성공적으로 처리되었습니다."),
+	CREATED(HttpStatus.CREATED, "새로운 리소스가 생성되었습니다.");
+
+	private final HttpStatus status;
+	private final String message;
+
+	@Override
+	public HttpStatus getStatus() {
+		return this.status;
+	}
+
+	@Override
+	public String getMessage() {
+		return this.message;
+	}
+
+}
